Separate tests for PBMessageWriter and PBFieldWriter
